Major refactor

+++ src/watchdog.c
@@ -2,12 +2,21 @@
 #include <string.h>
 #include <stdlib.h>
 #include <stdio.h>
+#include <sys/resource.h>
 #include <sys/mman.h>
 #include <sys/ptrace.h>
 #include <sys/wait.h>
+#include <sys/types.h>
+#include <sys/stat.h>
 #include <time.h>
 #include <pthread.h>
 #include <errno.h>
+#include <unistd.h>
+#include <math.h>
+
+#ifdef __linux__
+#include <sys/prctl.h>
+#endif
 
 #include "watchdog.h"
 #include "util.h"
@@ -18,12 +27,16 @@ int WatchDogRun(int argc, char **argv);
 
 int
 WatchDogInit(
-    void
+    void (*callback_on_stall)(void)
     )
 {
     pthread_mutexattr_t attrmutex;
     pthread_condattr_t attrcond;
     int status;
+
+    if(!ASSERT(callback_on_stall))
+    {   return EXIT_FAILURE;
+    }
 
     watchdog = mmap(NULL, sizeof(WatchDog), PROT_READ|PROT_WRITE, MAP_SHARED|MAP_ANONYMOUS, -1, 0);
 
@@ -49,6 +62,9 @@ WatchDogInit(
 
     pthread_condattr_setpshared(&attrcond, PTHREAD_PROCESS_SHARED);
 
+    /* cond doesnt support CLOCK_MONOTONIC_COARSE cor some reason -\(!-!)/- */
+    pthread_condattr_setclock(&attrcond, CLOCK_MONOTONIC);
+
     status = pthread_mutex_init(&watchdog->mutex, &attrmutex);
 
     if(status)
@@ -61,6 +77,7 @@ WatchDogInit(
     { 	goto UNMUTEX;
     }
 
+    watchdog->callback = callback_on_stall;
 
     return EXIT_SUCCESS;
 UNMUTEX:
@@ -88,13 +105,16 @@ WatchDogDestroy(
 int
 WatchDogStart(
     int argc,
-    char **argv
+    char **argv,
+    void (*callback_on_stall)(void)
     )
 {
     pid_t pid;
+    struct rlimit rl;
+    int fd;
 
     /* watchdog failed to start */
-    if(WatchDogInit() != EXIT_SUCCESS)
+    if(WatchDogInit(callback_on_stall) != EXIT_SUCCESS)
     {   return EXIT_FAILURE;
     }
 
@@ -106,15 +126,44 @@ WatchDogStart(
         case -1:
             return EXIT_FAILURE;
         case 0:
+            if(setsid() < 0)
+            {   exit(EXIT_FAILURE);
+            }
+
+            chdir("/");
+            umask(0);
+
+            getrlimit(RLIMIT_NOFILE, &rl);
+
+            /* stdin, stdout, stderr */
+            enum { IMPORTANT_FILE_DESCRIPTORS = 2 };
+
+            for(fd = 0; fd < rl.rlim_max; ++fd)
+            {   
+                if(fd > IMPORTANT_FILE_DESCRIPTORS)
+                {   close(fd);
+                }
+            }
+
+            #ifdef __linux__
+                prctl(PR_SET_NAME, "vox-wm-watchdog", 0, 0, 0);
+            #endif
             WatchDogRun(argc, argv);
-	    exit(EXIT_SUCCESS);
+	        exit(EXIT_SUCCESS);
             break;
         default:
             pthread_mutex_lock(&watchdog->mutex);
             watchdog->child = getpid();
-            watchdog->last_alive = time(NULL);
+
+            #if __linux__
+                clock_gettime(CLOCK_MONOTONIC_COARSE, &watchdog->last_alive);
+            #else
+                clock_gettime(CLOCK_MONOTONIC, &watchdog->last_alive);
+            #endif
+
             pthread_mutex_unlock(&watchdog->mutex);
             Debug0("WatchDog Succesfully Started!");
+            break;
     }
 
     return EXIT_SUCCESS;
@@ -146,18 +195,26 @@ WatchDogRequestExit(
 }
 
 void
-WatchDogPingAlive(
+WatchDogRespond(
     void
 )
 {
     if(!watchdog)
     {   return;
     }
+
     pthread_mutex_lock(&watchdog->mutex);
-
-    watchdog->last_alive = time(NULL);
-
+    watchdog->restart_timer = 1;
+    Debug0("shi");
     pthread_mutex_unlock(&watchdog->mutex);
+}
+
+static double
+timespecdiff(struct timespec *start, struct timespec *end)
+{
+    double sec = (double)(end->tv_sec - start->tv_sec);
+    double nsec = (double)(end->tv_nsec - start->tv_nsec) / 1000000000.0;
+    return sec + nsec;
 }
 
 int
@@ -177,24 +234,32 @@ WatchDogRun(
 
     /* due to speed, and blah blah blah we dont want this to hang or whatever... */
     const unsigned int SECONDS = 1;
-    const unsigned int MAX_TIMEOUT = 7;
+    const unsigned int MAX_TIMEOUT = 15;
     struct timespec PING_TIME;
 
     bool child_dead = false;
 
     while(watchdog->running && !watchdog->die)
     {
-	status = clock_gettime(CLOCK_REALTIME, &PING_TIME);
-	PING_TIME.tv_sec += SECONDS;
+        #if __linux__
+            clock_gettime(CLOCK_MONOTONIC_COARSE, &PING_TIME);
+        #else
+            clock_gettime(CLOCK_MONOTONIC, &PING_TIME);
+        #endif
+
+        PING_TIME.tv_sec += SECONDS;
         /* wait... */
         status = pthread_cond_timedwait(&watchdog->cond, &watchdog->mutex, &PING_TIME);
 
-	/* ok were kinda fucked here... 
-	 * This is here cause the above immediatly fails since the time is in the past....
-	 */
-	if(status == -1)
-	{   sleep(SECONDS);
-	}
+        /* ok were kinda fucked here... 
+         * This is here cause the above immediatly fails since the time is in the past....
+         */
+        if(status == -1)
+        {   
+            pthread_mutex_unlock(&watchdog->mutex);
+            sleep(SECONDS);
+            pthread_mutex_lock(&watchdog->mutex);
+        }
 
         /* is the process alive? */
         if(kill(watchdog->child, 0) != EXIT_SUCCESS)
@@ -209,18 +274,56 @@ WatchDogRun(
             }
         }
 
-        time_t now = time(NULL);
-	Debug("%ld", now - watchdog->last_alive);
+        struct timespec now;
 
-        /* 7 seconds is more than enough. */
-        if (now - watchdog->last_alive > MAX_TIMEOUT)
+        #if __linux__
+            status = clock_gettime(CLOCK_MONOTONIC_COARSE, &now);
+        #else
+            status = clock_gettime(CLOCK_MONOTONIC, &now);
+        #endif
+
+        /* we are beyond fucked */
+        if(status == -1)
+        {   
+            watchdog->running = 0;
+            break;
+        }
+
+        double timeLeft = MAX_TIMEOUT - timespecdiff(&watchdog->last_alive, &now);
+
+        #if DEBUG
+            fprintf(stderr, "\033[2K\r[src/watchdog.c:405] by doomTimer(): %.0f ", timeLeft);
+            fflush(stderr);
+        #endif
+
+        if(watchdog->restart_timer)
+        {
+            watchdog->restart_timer = 0;
+            /* this might drift and be a problem for long sessions... Maybe? */
+            watchdog->last_alive.tv_sec += MAX_TIMEOUT - (time_t)ceil(timeLeft);
+        }
+        else
+        {
+            const float CHECK_THRESHOLD = .75;
+
+            if(MAX_TIMEOUT * CHECK_THRESHOLD > timeLeft)
+            {   
+                watchdog->callback();
+                Debug0("ru");
+            }
+        }
+
+        /* have we run out of time? */
+        if (timeLeft < 0)
         {   
             kill(watchdog->child, SIGKILL);
-	    #ifdef DEBUG
-		kill(watchdog->child, SIGABRT);
-	    #else
-		kill(watchdog->child, SIGKILL);
-	    #endif
+
+            #ifdef DEBUG
+                kill(watchdog->child, SIGABRT);
+            #else
+                kill(watchdog->child, SIGKILL);
+            #endif
+
             watchdog->running = 0;
             /* Process might hang for a undefined length of time so this just forces a restart regardless. */
             watchdog->restart = 1;
@@ -264,7 +367,7 @@ WatchDogRun(
     if(watchdog->die)
     {   
         pthread_mutex_unlock(&watchdog->mutex);
-	_exit(EXIT_SUCCESS);
+        _exit(EXIT_SUCCESS);
         return EXIT_SUCCESS;
     }
 
@@ -282,14 +385,15 @@ WatchDogRun(
         /* This should be unreachable but could maybe happen. */
         if(unlikely(!child_dead))
         {   	
-	    #ifdef DEBUG
-		kill(watchdog->child, SIGABRT);
-	    #else
-		kill(watchdog->child, SIGKILL);
-	    #endif
+            #ifdef DEBUG
+                kill(watchdog->child, SIGABRT);
+            #else
+                kill(watchdog->child, SIGKILL);
+            #endif
         }
 
         Debug0("Watchdog Restarting Process...");
+
         #ifdef __linux__
             char *path = "/proc/self/exe";
             execv(path, argv);
@@ -303,11 +407,10 @@ WatchDogRun(
     if(likely(child_dead))
     {   
         Debug0("Exiting, child dead?? FIXME.");
-	_exit(EXIT_SUCCESS);
+        _exit(EXIT_SUCCESS);
     }
 
-
-    ASSERT(false);
+    (void)ASSERT(false);
 
     return EXIT_FAILURE;
 }
